+++ src/move.h
@@ -14,8 +14,8 @@ void move_up(int count);
 void move_down(int count);
 void move_bof(void);
 void move_eof(void);
-void move_to_line(View *v, int line);
-void move_to_column(View *v, int column);
+void move_to_line(View *v, size_t line);
+void move_to_column(View *v, size_t column);
 
 size_t word_fwd(BlockIter *bi, bool skip_non_word);
 size_t word_bwd(BlockIter *bi, bool skip_non_word);
